Implementado o basesql e criado teste inicial de adicionar pedido

+++ projeto-pizzaria/projeto-pizzaria.Domain/Interfaces/IRepositorio.cs
@@ -9,7 +9,7 @@ namespace projeto_pizzaria.Domain.Interfaces
 {
     public interface IRepositorio<T> where T : Entidade
     {
-        int Adicionar(T entidade);
+        long Adicionar(T entidade);
         void Editar(T entidade);
         void Excluir(T entidade);
         List<T> GetAll();
